change: manage session with status instead of pending value

+++ sql/sessions.sql
@@ -4,7 +4,7 @@ create table sessions
     `expires_at`    bigint       not null,
     `user_id`       int          not null,
     `ip`            varchar(255) not null,
-    `pending`       tinyint(1)   DEFAULT 0,
+    `pending`       int   DEFAULT 0,
     constraint sessions_pk
         primary key (`state`)
 );
